# Add diagnostic metrics/activities