--- a/build/System.Memory.props
+++ b/build/System.Memory.props
@@ -1,6 +1,0 @@
-<Project DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <!-- '!NET6_0_OR_GREATER' equivalent -->
-  <ItemGroup Condition="!('$(TargetFrameworkIdentifier)' == '.NETCoreApp' AND $([MSBuild]::VersionGreaterThanOrEquals($(TargetFrameworkVersion), '6.0')))">
-    <PackageReference Include="System.Memory" Version="4.5.3" />
-  </ItemGroup>
-</Project>