--- a/src/Avalonia.Base/Diagnostics/Diagnostic.Activities.cs
+++ b/src/Avalonia.Base/Diagnostics/Diagnostic.Activities.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Diagnostics;
+
+// ReSharper disable ExplicitCallerInfoArgument
+
+namespace Avalonia.Diagnostics;
+
+internal static partial class Diagnostic
+{
+    private static readonly ActivitySource s_diagnostic = new("Avalonia.Diagnostic.Source");
+}