--- a/src/Avalonia.Base/Diagnostics/Diagnostic.Activities.cs
+++ b/src/Avalonia.Base/Diagnostics/Diagnostic.Activities.cs
@@ -8,4 +8,8 @@ namespace Avalonia.Diagnostics;
 internal static partial class Diagnostic
 {
     private static readonly ActivitySource s_diagnostic = new("Avalonia.Diagnostic.Source");
+
+    public static Activity? FindingResourceActivity() => s_diagnostic
+        .StartActivity("Avalonia.Controls.ResourceNode.FindResource");
+
 }