--- a/src/Avalonia.Base/Diagnostics/Diagnostic.Activities.cs
+++ b/src/Avalonia.Base/Diagnostics/Diagnostic.Activities.cs
@@ -9,7 +9,13 @@ internal static partial class Diagnostic
 {
     private static readonly ActivitySource s_diagnostic = new("Avalonia.Diagnostic.Source");
 
+    public static Activity? AttachingStyleActivity() => s_diagnostic
+        .StartActivity("Avalonia.Styling.Style.Attach");
+
     public static Activity? FindingResourceActivity() => s_diagnostic
         .StartActivity("Avalonia.Controls.ResourceNode.FindResource");
+
+    public static Activity? EvaluatingStyleActivator() => s_diagnostic
+        .StartActivity("Avalonia.Styling.Activators.StyleActivatorBase.EvaluateIsActive");
 
 }