--- a/src/Avalonia.Base/Diagnostics/Diagnostic.Consts.cs
+++ b/src/Avalonia.Base/Diagnostics/Diagnostic.Consts.cs
@@ -20,5 +20,32 @@ internal static partial class Diagnostic
         public const string LayoutRenderPassDescription = "Duration of render recording pass on UI thread";
         public const string LayoutInputPassName = "avalonia.ui.input.time";
         public const string LayoutInputPassDescription = "Duration of input processing on UI thread";
+
+        public const string TotalEventHandleCountName = "avalonia.ui.event.handler.count";
+        public const string TotalEventHandleCountDescription = "Number of event handlers currently registered in the application";
+        public const string TotalEventHandleCountUnit = "{handler}";
+        public const string TotalVisualCountName = "avalonia.ui.visual.count";
+        public const string TotalVisualCountDescription = "Number of visual elements currently present in the visual tree";
+        public const string TotalVisualCountUnit = "{visual}";
+        public const string TotalDispatcherTimerCountName = "avalonia.ui.dispatcher.timer.count";
+        public const string TotalDispatcherTimerCountDescription = "Number of active dispatcher timers in the application";
+        public const string TotalDispatcherTimerCountUnit = "{timer}";
+    }
+
+    public static class Tags
+    {
+        public const string Style = nameof(Style);
+        public const string SelectorResult = nameof(SelectorResult);
+
+        public const string Key = nameof(Key);
+        public const string ThemeVariant = nameof(ThemeVariant);
+        public const string Result = nameof(Result);
+
+        public const string Activator = nameof(Activator);
+        public const string IsActive = nameof(IsActive);
+        public const string Selector = nameof(Selector);
+        public const string Control = nameof(Control);
+
+        public const string RoutedEvent = nameof(RoutedEvent);
     }
 }