--- a/src/Avalonia.Base/Diagnostics/Diagnostic.Consts.cs
+++ b/src/Avalonia.Base/Diagnostics/Diagnostic.Consts.cs
@@ -1,7 +1,12 @@
+using System;
+
 namespace Avalonia.Diagnostics;
 
 internal static partial class Diagnostic
 {
+    public static bool IsEnabled { get; } = InitializeIsEnabled();
+    private static bool InitializeIsEnabled() => AppContext.TryGetSwitch("Avalonia.Diagnostics.Diagnostic.IsEnabled", out var isEnabled) && isEnabled;
+
     public static class Meters
     {
         public const string SecondsUnit = "s";