--- a/src/Avalonia.Base/Diagnostics/Diagnostic.Metrics.cs
+++ b/src/Avalonia.Base/Diagnostics/Diagnostic.Metrics.cs
@@ -8,54 +8,52 @@ namespace Avalonia.Diagnostics;
 
 internal static partial class Diagnostic
 {
-    private static readonly Meter s_meter;
+    private static Histogram<double>? s_compositorRender;
+    private static Histogram<double>? s_compositorUpdate;
+    private static Histogram<double>? s_layoutMeasure;
+    private static Histogram<double>? s_layoutArrange;
+    private static Histogram<double>? s_layoutRender;
+    private static Histogram<double>? s_layoutInput;
 
-    private static readonly Histogram<double> s_compositorRender;
-    private static readonly Histogram<double> s_compositorUpdate;
-
-    private static readonly Histogram<double> s_layoutMeasure;
-    private static readonly Histogram<double> s_layoutArrange;
-    private static readonly Histogram<double> s_layoutRender;
-    private static readonly Histogram<double> s_layoutInput;
-
-    static Diagnostic()
+    public static void InitMetrics()
     {
-        s_meter = new Meter("Avalonia.Diagnostic.Meter");
-        s_compositorRender = s_meter.CreateHistogram<double>(
+        // Metrics
+        var meter = new Meter("Avalonia.Diagnostic.Meter");
+        s_compositorRender = meter.CreateHistogram<double>(
             Meters.CompositorRenderPassName,
             Meters.MillisecondsUnit,
             Meters.CompositorRenderPassDescription);
-        s_compositorUpdate = s_meter.CreateHistogram<double>(
+        s_compositorUpdate = meter.CreateHistogram<double>(
             Meters.CompositorUpdatePassName,
             Meters.MillisecondsUnit,
             Meters.CompositorUpdatePassDescription);
-        s_layoutMeasure = s_meter.CreateHistogram<double>(
+        s_layoutMeasure = meter.CreateHistogram<double>(
             Meters.LayoutMeasurePassName,
             Meters.MillisecondsUnit,
             Meters.LayoutMeasurePassDescription);
-        s_layoutArrange = s_meter.CreateHistogram<double>(
+        s_layoutArrange = meter.CreateHistogram<double>(
             Meters.LayoutArrangePassName,
             Meters.MillisecondsUnit,
             Meters.LayoutArrangePassDescription);
-        s_layoutRender = s_meter.CreateHistogram<double>(
+        s_layoutRender = meter.CreateHistogram<double>(
             Meters.LayoutRenderPassName,
             Meters.MillisecondsUnit,
             Meters.LayoutRenderPassDescription);
-        s_layoutInput = s_meter.CreateHistogram<double>(
+        s_layoutInput = meter.CreateHistogram<double>(
             Meters.LayoutInputPassName,
             Meters.MillisecondsUnit,
             Meters.LayoutInputPassDescription);
-        s_meter.CreateObservableUpDownCounter(
+        meter.CreateObservableUpDownCounter(
             Meters.TotalEventHandleCountName,
             () => Interactive.TotalHandlersCount,
             Meters.TotalEventHandleCountUnit,
             Meters.TotalEventHandleCountDescription);
-        s_meter.CreateObservableUpDownCounter(
+        meter.CreateObservableUpDownCounter(
             Meters.TotalVisualCountName,
             () => Visual.RootedVisualChildrenCount,
             Meters.TotalVisualCountUnit,
             Meters.TotalVisualCountDescription);
-        s_meter.CreateObservableUpDownCounter(
+        meter.CreateObservableUpDownCounter(
             Meters.TotalDispatcherTimerCountName,
             () => DispatcherTimer.ActiveTimersCount,
             Meters.TotalDispatcherTimerCountUnit,
@@ -69,8 +67,8 @@ internal static partial class Diagnostic
     public static HistogramReportDisposable BeginLayoutInputPass() => Begin(s_layoutInput);
     public static HistogramReportDisposable BeginLayoutRenderPass() => Begin(s_layoutRender);
 
-    private static HistogramReportDisposable Begin(Histogram<double> histogram) => IsEnabled ? new(histogram) : default;
-    
+    private static HistogramReportDisposable Begin(Histogram<double>? histogram) => histogram is not null ? new(histogram) : default;
+
     internal readonly ref struct HistogramReportDisposable
     {
         private readonly Histogram<double> _histogram;