… src/Avalonia.Base/Diagnostics/Diagnostic.cs
… src/Avalonia.Base/Interactivity/EventRoute.cs
… src/Avalonia.Base/Interactivity/Interactive.cs
… src/Avalonia.Base/Interactivity/RoutedEvent.cs
… src/Avalonia.Base/Layout/LayoutManager.cs
… src/Avalonia.Base/Layout/Layoutable.cs
… src/Avalonia.Base/Rendering/Composition/CompositingRenderer.cs
… src/Avalonia.Base/Rendering/Composition/Server/ServerCompositionTarget.cs
… src/Avalonia.Base/Styling/ControlTheme.cs
… src/Avalonia.Base/Styling/Style.cs
… src/Avalonia.Base/Styling/StyleInstance.cs
… src/Avalonia.Base/Threading/DispatcherTimer.cs
… src/Avalonia.Base/Utilities/StopwatchHelper.cs
… src/Avalonia.Base/Visual.cs
… src/Avalonia.Controls/TopLevel.cs
… src/Markup/Avalonia.Markup.Xaml/MarkupExtensions/StaticResourceExtension.cs
… src/Markup/Avalonia.Markup/Avalonia.Markup.csproj
… tests/Avalonia.Controls.DataGrid.UnitTests/Avalonia.Controls.DataGrid.UnitTests.csproj